+++ packages/decap-cms-backend-aws-cognito-github-proxy/src/implementation.tsx
@@ -1,11 +1,10 @@
 import * as React from 'react';
 import { GitHubBackend } from 'decap-cms-backend-github';
-
-import AuthenticationPage from './AuthenticationPage';
+import { PKCEAuthenticationPage } from 'decap-cms-ui-auth';
 
 import type { GitHubUser } from 'decap-cms-backend-github/src/implementation';
 import type { Config } from 'decap-cms-lib-util/src';
-import type { Octokit } from '@octokit/rest';
+import type { Endpoints } from '@octokit/types';
 
 export default class AwsCognitoGitHubProxyBackend extends GitHubBackend {
   constructor(config: Config, options = {}) {
@@ -16,9 +15,10 @@ export default class AwsCognitoGitHubProxyBackend extends GitHubBackend {
   }
 
   authComponent() {
-    const wrappedAuthenticationPage = (props: Record<string, unknown>) => (
-      <AuthenticationPage {...props} backend={this} />
-    );
+    const wrappedAuthenticationPage = (props: Record<string, unknown>) => {
+      const allProps = { ...props, backend: this };
+      return <PKCEAuthenticationPage {...allProps} />;
+    };
     wrappedAuthenticationPage.displayName = 'AuthenticationPage';
     return wrappedAuthenticationPage;
   }
@@ -46,7 +46,9 @@ export default class AwsCognitoGitHubProxyBackend extends GitHubBackend {
     return this._currentUserPromise;
   }
 
-  async getPullRequestAuthor(pullRequest: Octokit.PullsListResponseItem) {
+  async getPullRequestAuthor(
+    pullRequest: Endpoints['GET /repos/{owner}/{repo}/pulls']['response']['data'][0],
+  ) {
     return pullRequest.user?.login;
   }
 }