+++ packages/decap-cms-backend-azure/CHANGELOG.md
@@ -3,6 +3,26 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [3.4.0](https://github.com/decaporg/decap-cms/compare/decap-cms-backend-azure@3.3.0...decap-cms-backend-azure@3.4.0) (2025-12-18)
+
+### Bug Fixes
+
+- handle potential null labels in pull request filtering ([#7678](https://github.com/decaporg/decap-cms/issues/7678)) ([fce2a69](https://github.com/decaporg/decap-cms/commit/fce2a69a217298771fa4f01b985de8116ed94e1a))
+
+# [3.3.0](https://github.com/decaporg/decap-cms/compare/decap-cms-backend-azure@3.2.1...decap-cms-backend-azure@3.3.0) (2025-07-15)
+
+### Features
+
+- add logo to header ([#7487](https://github.com/decaporg/decap-cms/issues/7487)) ([b540ace](https://github.com/decaporg/decap-cms/commit/b540acec943eb231df6aac7b1d515d9b4b84fa5d))
+
+## [3.2.1](https://github.com/decaporg/decap-cms/compare/decap-cms-backend-azure@3.2.0...decap-cms-backend-azure@3.2.1) (2025-07-10)
+
+**Note:** Version bump only for package decap-cms-backend-azure
+
+# [3.2.0](https://github.com/decaporg/decap-cms/compare/decap-cms-backend-azure@3.1.3...decap-cms-backend-azure@3.2.0) (2025-06-26)
+
+**Note:** Version bump only for package decap-cms-backend-azure
+
 ## [3.1.3](https://github.com/decaporg/decap-cms/compare/decap-cms-backend-azure@3.1.2...decap-cms-backend-azure@3.1.3) (2024-08-13)
 
 ### Reverts